finalizado

+++ src/games/platform/models/Discounts.java
@@ -6,10 +6,10 @@ public class Discounts {
 
     private int id;
     private int game_id;
-    private float price_discount;
+    private int price_discount;
     private Date expire_date;
 
-    public Discounts(int id, int game_id, float price_discount, Date expire_date) {
+    public Discounts(int id, int game_id, int price_discount, Date expire_date) {
         this.id = id;
         this.game_id = game_id;
         this.price_discount = price_discount;
@@ -24,7 +24,7 @@ public class Discounts {
         return game_id;
     }
 
-    public float getPrice_discount() {
+    public int getPrice_discount() {
         return price_discount;
     }
 
@@ -40,7 +40,7 @@ public class Discounts {
         this.game_id = game_id;
     }
 
-    public void setPrice_discount(float price_discount) {
+    public void setPrice_discount(int price_discount) {
         this.price_discount = price_discount;
     }
